+++ Data/Scripts/WeaponCore/Definitions/WeaponDefinitions.cs
@@ -654,7 +654,7 @@ namespace WeaponCore.Support
         {
             [ProtoMember(1)] internal float Scale;
             [ProtoMember(2)] internal float MaxDistance;
-            [ProtoMember(3)] internal uint MaxDuration;
+            [ProtoMember(3)] internal float MaxDuration;
             [ProtoMember(4)] internal bool Loop;
             [ProtoMember(5)] internal bool Restart;
             [ProtoMember(6)] internal float HitPlayChance;
